LayoutTests/storage/indexeddb/readonly-properties-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test IndexedDB readonly properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
transaction = event.target.result;
Deleted all object stores.
objectStore = db.createObjectStore('foo');
trying to set readonly property objectStore.transaction
objectStore.transaction = this
PASS objectStore.transaction is still [object IDBTransaction]
PASS successfullyParsed is true

TEST COMPLETE
